set_ccid_descriptor(): reset dwSlotStatus to IFD_ICC_PRESENT for the other slots of a multislot reader (like a GemCore SIM Pro).
This is needed because the state of dwSlotStatus may have already been
changed to IFD_ICC_NOT_PRESENT (by the polling thread) when the second
slot is created. The polling thread of the second slot would then never
check for a card since this check is only done once. Slots are SAMs and
the card is always present or absent.

The problem was already dealt with on USB from the beginning but not on
serial.

Thanks to Emmanuel Deloget for the patch.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@4521 0ce88b0d-b2fd-0310-8134-9614164e65ea