CreateChannelByNameOrChannel(): Do not change the default timeout for the first call to CmdGetSlotStatus().
In revision 4938 the timeout is changed from default to 100 ms to avoid
a 2 seconds delay if the reader needs to resynchronize.

In revision 5389 the timeout was changed from 100 ms to 1000 ms because
it was too sort for the Kobil KAAN Base reader.

The idea now is to use the default value (DEFAULT_COM_READ_TIMEOUT
(3*1000)) for the first CmdGetSlotStatus() call. And then use a short
value (100 ms) for the 2 next calls.

The reader should have enough time to start during the 3 first seconds.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6396 0ce88b0d-b2fd-0310-8134-9614164e65ea