CreateChannelByNameOrChannel(): reset ccid_descriptor->readTimeout in all cases
The value was not restored in the error case. So it was not a bug. But
the Coverity tool complained:

CID 1090609 (#1 of 1): Failure to restore non-local value
(MISSING_RESTORE)20. end_of_scope: Value of non-local
"ccid_descriptor->readTimeout" that was saved in "oldReadTimeout" is not
restored as it was along other paths.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6753 0ce88b0d-b2fd-0310-8134-9614164e65ea