Add a timeout parameter to CmdEscape()
Not all commands need the same timeout. For example the CmdEscape() used
in ccid_serial.c to check if the reader is alive should have a short
timeout.
The timeout was forced to 30 seconds inside CmdEscape() so it was not
possible to use a different value. The patch fixes this.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6783 0ce88b0d-b2fd-0310-8134-9614164e65ea