IFDHControl(): report FEATURE_IFD_PIN_PROPERTIES only for pinpad readers
Some (bogus) applications use the presence of FEATURE_IFD_PIN_PROPERTIES
to know if the reader is a pinpad reader or not.
The application should use FEATURE_VERIFY_PIN_DIRECT and
FEATURE_MODIFY_PIN_DIRECT instead.

Reporting FEATURE_IFD_PIN_PROPERTIES for a non-pinpad reader is useless
so it should not harm to remove it from CM_IOCTL_GET_FEATURE_REQUEST.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6790 0ce88b0d-b2fd-0310-8134-9614164e65ea