Generalize the management of readers with bDeviceClass = 0xFF
Some early Gemalto Ezio CB+ readers have bDeviceClass, bDeviceSubClass and
bDeviceProtocol set to 0xFF (proprietary) instead of 0x00.

So on Mac OS X the reader configuration is not done by the OS/kernel and we do
it ourself.

The code is no more limited to the GEMALTO_EZIO_CBP (0x08E6/0x34C3) but to any
reader with the problem, like the Todos Argos Mini II CCID.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6793 0ce88b0d-b2fd-0310-8134-9614164e65ea