Fix calculation of the command length
Since pcsc-lite 1.8.9 (October 2013) the structures PIN_MODIFY_STRUCTURE
and PIN_VERIFY_STRUCTURE are different and now use a C99 flexible array
member when available for abData field. uint8_t abData[];

The the sizeof(PIN_VERIFY_STRUCTURE) is one byte smaller and the
calculation of the complete command is now simpler.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6809 0ce88b0d-b2fd-0310-8134-9614164e65ea