get_end_points(): be more strict for bInterfaceClass = 255
If the bInterfaceClass is 255 (proprietary) and then not, but possibly,
a CCID interface (on an old reader) we also check the extra_length is
valid for a CCID device.

The problem was with a composite device with 2 interfaces: one non-CCID
interface with bInterfaceClass = 255 and a CCID interface with
bInterfaceClass = 11. The first non-CCID interface must be skipped.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6828 0ce88b0d-b2fd-0310-8134-9614164e65ea