t1_transceive(): check retries is is positive
The test was "retries == 0" but it may happen that retries becomes
negative. Then the counter will be decreasing "indefinitely".

The new test is "retries <= 0" to also get the negative values and
resync if needed.

The problem was detected by a T=1 card on a TPDU pinpad reader with a
VERIFY command that send a WTX request before returning the SW code.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6868 0ce88b0d-b2fd-0310-8134-9614164e65ea