The O2 Micro Oz776 reader only supports 9600 bps
The reader returns bNumDataRatesSupported: 0 and a list with only 1
value: 9600 bps.
bNumDataRatesSupported: 0 should indicate that all the speeds between
dwDataRate: 9600 bps and dwMaxDataRate: 307200 bps are available. But it
looks like this is not the case.

The driver behavior changed with revision 6801 "Do not get the data
rates if bNumDataRatesSupported = 0".

This patch just set dwMaxDataRate to 9600 so only this speed is used.
The O2 Micro Oz776 reader was already in the unsupported list.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6885 0ce88b0d-b2fd-0310-8134-9614164e65ea