Remove bogus treatment for CARDMAN3121
The code was not used by the old CARDMAN3121 but had a side effect for a
new version of the Cardman 3121 with Product ID increased by 1.

This code was present since the first version of the file in subversion
or CVS (September 2003). I do not remember why I added this code, more
than 10 years ago.

Thanks to Maximilian Stein for the bug report
http://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20140714/000073.html


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6940 0ce88b0d-b2fd-0310-8134-9614164e65ea