Enable effect of --enable-embedded on hex dumps
Use LogXxd() instead of log_xxd() in the definition of DEBUG_XXD and
DEBUG_INFO_XXD.

LogXxd() is defined as nothing (and removed by the compiler) if NO_LOG
is enabled (using --enable-embedded). So the header text of DEBUG_XXD()
is no more stored in the .rodata section of the binary file.

'objdump -h --section=.rodata libccid.so' shows a gain of 38 bytes.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@6967 0ce88b0d-b2fd-0310-8134-9614164e65ea