Add CmdEscapeCheck()
CmdEscapeCheck() is CmdEscape() with an extra "mayfail" parameter.

If mayfail is TRUE then the error is logged as PCSC_LOG_INFO instead of
PCSC_LOG_ERROR.

This is usefull in cases we know the command may fail. So the error is
expected and not really an error.

See Alioth bug [#314871] "commands.c:979:CmdEscape error on byte 10"           
message
https://alioth.debian.org/tracker/index.php?func=detail&aid=314871&group_id=3010+++5&atid=410086


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@7039 0ce88b0d-b2fd-0310-8134-9614164e65ea