OpenUSBByName(): free the device list in case of error
If no reader is found the code should release the device list in any
case.

Either the code will try again to find the reader (in case of Mac OS X)
or return an error code.

The list of USB devices shall always be released before returning from the
function (in error or normal cases).