CreateChannel(): cleanup in case of error
If the initialisation fails after the OpenPortByName() (during the check
and hack phases) then we need to correctly close the device.

The problem occured on OS X with a Gemalto Pinpad reader.
The ccid_open_hack_post() failed and the port was not closed.
usbDevice[reader_index].dev_handle was not set to NULL and the next
reader plug (any reader) failed because the index was already in use.