Set timeout to 90 sec for PPDU commands
PPDU (Pseudo APDU) commands are described in PC/SC v2 part 10
supplement document. They are defined as CLA=0xFF, INS=0xC2, P1=0x01.
When a PPDU is sent through SCardTransmit() the communication timeout is
changed to 90 seconds (the same as for a SCardControl() command).

This change allows the use of a Secure Verify command sent as a PPDU
through SCardTransmit(). The calculated timeout of 2.85 seconds was too
short in this case.

Fix alioth bug [#315217] "Cherry KC 1000 SC direct PIN modification
timeout Pseudo APDU"
https://alioth.debian.org/tracker/?func=detail&atid=410085&aid=315217&group_id=30105

Thanks to Roman Pasechnik for the bug report.