OpenUSBByName(): use device bus and address only when available
Use the device_bus and device_addr only if they have a value i.e. when
libudev is used by pcscd.

If pcsc-lite uses libusb instead of libudev for the hotplug mechanism
the driver is called with device="usb:08e6/3437:libusb-1.0:1:10:0" and
then device_bus and device_addr are set to 0.  In this case the driver
must not uses these values.
The patch in 8b9f6f33b16e808a5769dbc4193e5f107a09a693 was not complete.

Thanks to Luka Logar for the bug report and patch.