OpenUSBByName(): read Info.plist only once
On Mac OS X it is possible to ask libusb to rescan the USB bus waiting
for a device o be available.
See commit 690e90f476c3815c31874e523ba017f7f1d372cd

The problem was that the code to parse the driver Info.plist was also
re-executed. This code is now moved earlier and is executed only once
even on Mac OS X.