Patch for Microchip SEC1110 on Mac OS X
The Microchip SEC1110 is bogus and does not report card status correctly
with PC_to_RDR_GetSlotStatus if the Interrupt pipe is not read.

The problem is that Mac OS X still does not use the
TAG_IFD_POLLING_THREAD_WITH_TIMEOUT and then the Interrupt pipe to
detect card events. Only IFDHICCPresence(), and then GetSlotStatus, is
used.

The solution is to empty the interrupt pipe by doing a very short read
(10 ms) before each call to PC_to_RDR_GetSlotStatus.