Patch for Cherry KC 1000 SC
The Cherry KC 1000 SC is (by default) a "Short APDU level exchange" CCID
reader. But firmware 1.00 is bogus.
With a T=1 card and case 2 APDU (data from card to host) the maximum
size returned by the reader is 128 byes. The reader is then using
chaining as with extended APDU.

In order to limit the impacts on the driver I just change the dwFeatures
to declare the reader as a "Short and Extended APDU level exchange" CCID
reader to use the chaining even for short APDU.

Thanks to ZF Friedrichshafen AG for the patch.