CreateChannel: allow to use a non default timeout
By default the read timeout is set to 100 ms for the double call of
CmdGetSlotStatus() in CreateChannelByNameOrChannel() use to resync the
USB toggle bits.
This is a short delay so that a failing reader will not block for too
long. The defaut timeout value is 3 seconds so a total of 6 seconds
would be lost is the reader is not responding.

Unfortunately some readers (in particular contactless readers) are very
long to anwser to the PC_to_RDR_GetSlotStatus CCID command. I guess it
is to be sure no contactless card is present in the field.

It is now possible to change the 100 ms timeout to something else for a
specific reader by adding code in ccid_open_hack_pre() and
ccid_open_hack_post().