IFDHGetCapabilities: fix SCARD_ATTR_VENDOR_IFD_SERIAL_NO size
We use strlcpy() to copy the IFD (reader) serial number string. So we
must must take into account the terminating NUL byte.

Without the patch the size was "correctly" reported on the first call
(without including the final NUL byte). So the application allocates a
buffer for this size and the next strlcpy() tuncates the string to add
the NUL byte.

Thanks to Daniel Kucera for the bug report.