log_msg(): log up to 3 lines of 80 characters
The log of the CCID driver license was truncated. It is now possible to
log up to 3 * 80 = 240 characters instead of 2 * 80 = 160.

The problem is only present on Mac OS X and other systems (Android)
where the log is not done by pcscd.
pcscd has a limit of 2048 bytes and does not have this problem.

Thanks to Maksim Ivanov for the bug report.
"[Pcsclite-muscle] Small fix suggestion for CCID (logging issue)"
https://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20160613/000634.html