Move Broadcom Corp 5880 0x5805 in unsupported
The Broadcom Corp 5880 with productID 0x5805 has some serious issues.

The contact interface of the reader is bogus.
Frames bigger than 64 bytes (wMaxPacketSize) fails on the contact reader
with:
00000006 commands.c:1740:CmdXfrBlockTPDU_T0() T=0: 5 bytes
00000011 -> 000000 6F 05 00 00 00 00 5A 00 00 00 80 34 01 00 00
00345316 <- 000000 80 02 01 00 00 00 5A 00 00 00 00 01 02 03 04 05 06
07 08 09 0A 0B 0C 0D 0E 0F 10 11 12 13 14 15 16 17 18 19 1A 1B 1C 1D 1E
1F 20 21 22 23 24 25 26 27 28 29 2A 2B 2C 2D 2E 2F 30 31 32 33 34 35
00000095 commands.c:1551:CCID_Receive() Can't read all data (54 out of
258 expected)

The reader _MAY_ work with a T=1 card since it is in TPDU.

This reader is present in Dell Latitude serie 5000 laptops.