PowerOn: change default voltage strategy
The CCID 1.1 indicates page 105 in ch. "9.3 Voltage management" that the
driver should use growing voltage when supported by the reader.

So the default is now to use the sequence 1.8V, 3V and 5V.

ISO 7816-3 is not clear about this point. No specific order is given in
ch. 6.2.4 "Class selection" of ISO/IEC 7816-3:2006(E).

This change should fix RedHat bug 1380615 "Coolkey cards do not work in
Fedora" and avoid the RedHat only patch described in RedHat bug 1033788
"Safenet SC650 smart cards are not detected"

The RedHat patch is
https://git.centos.org/blob/rpms!pcsc-lite-ccid/3dd85f55c8aacfa0704849e7b6cf46d4970b747b/SOURCES!ccid-1.4.10-voltage.patch