PowerOn: always try every possible voltage values
The default voltage value is 5V.
It is possible to change the driver configuration (ifdDriverOptions in
Info.plist) to start at 1.8V or 3V.

The algorithms was to try:
5V then fail (default)
3V then 5V then fail
1.8V then 3V then 5V then fail

Some cards do not support 5V (like the Coolkey cards, see "Coolkey cards
do not work in Fedora" https://bugzilla.redhat.com/show_bug.cgi?id=1380615)
It is possible to change the driver configuration to start at 3V or 1.8V
but then for cards that work only at 5V we will have 1 or 2 PowerOn
commands in error before the card returns the ATR.

I made some tests and the SCardConnect() call can take 0.110129 ms when
starting at 5V and 0.216057 ms when starting at 1.8V and errors for 1.8V
and 3V. That is "speed" factor of x1.96.

00000005 ifdhandler.c:1146:IFDHPowerICC() action: PowerUp,
usb:08e6/3437:libudev:0:/dev/bus/usb/001/041 (lun: 0)
00000007 -> 000000 62 00 00 00 00 00 08 03 00 00
00054697 <- 000000 80 00 00 00 00 00 08 41 FE 00
00000014 commands.c:249:CmdPowerOn Card absent or mute
00000005 commands.c:278:CmdPowerOn() Power up with 1.8V failed. Try with 3V.
00000006 -> 000000 62 00 00 00 00 00 09 02 00 00
00054723 <- 000000 80 00 00 00 00 00 09 41 FE 00
00000015 commands.c:249:CmdPowerOn Card absent or mute
00000004 commands.c:278:CmdPowerOn() Power up with 3V failed. Try with 5V.
00000006 -> 000000 62 00 00 00 00 00 0A 01 00 00
00037738 <- 000000 80 0C 00 00 00 00 0A 00 00 00 3B 29 00 80 72 A4 45 64 00 00 87 14

The idea is to change the algotithm to start at 5V as that is the
standard voltage for a smart card (according to my card collection) but
be able to continue with 1.8V and 3V is power on failed at 5V.

The default algorithm is now 5V then 1.8V then 3V then fail.

It is still possible to change the initial voltage. Now, in any case, all
the values are tried before failing.
Other possible values are:
1.8V then 3V then 5V then fail
3V then 5V then 1.8V then fail

Thanks to Robert Relyea for the patch, used by RedHat since 2013.
https://git.centos.org/commitdiff/rpms!pcsc-lite-ccid/3dd85f55c8aacfa0704849e7b6cf46d4970b747b