Negociate maximum baud rate when bNumDataRatesSupported = 0
From CCID 1.1 page 18:
" bNumDataRatesSupported
The number of data rates that are supported by the CCID.
If the value is 00h, all data rates between the default data rate
dwDataRate and the maximum data rate dwMaxDataRate are supported. "

Before, if the card baud rate was greater than the reader dwMaxDataRate
then the driver used the default (lowest) baud rate.

Now the driver tries to find the maximal baud rate supported by both the
card and the driver.

The algorithm was already used when the reader reports a baud rate list
but not when the reader reports bNumDataRatesSupported = 0.

Thanks to Abraham Macías Paredes for the bug report
"[Pcsclite-muscle] Bug in CCID library"
https://lists.alioth.debian.org/pipermail/pcsclite-muscle/Week-of-Mon-20161003/000729.html

Thanks to Godfrey Chung for the patch idea
"Fix the optimal baud rate problem in IFDHSetProtocolParameters()."
https://github.com/acshk/acsccid/commit/4a7631d014781d36a2d21210b0b8768c07ef8c6d