Fix non-pinpad HID global devices
The HID Aviator is not the only reader that declares pinpad support in
the USB descriptor but that has no physical keyboard to enter a PIN.
The USB descriptor is bogus.

The readers also support Extended APDU even if they declare Short APDU.

Bogus readers are:
- OMNIKEY Generic
- OMNIKEY 3121 or 3021 or 1021
- OMNIKEY 6121
- Cherry Smart Terminal xx44
- Fujitsu Smartcard Reader D323

Thanks to hidglobal.com for the patch.