Add ACS ACR33U to list of "should work" readers
This reader is a reader with a total of 5 slots (2x full size, 3x SIM size)
which makes it an attractive target for testing multi-slot CCID support.

The two external full-size slots work as expected, and report card
insert/removal status.  The three SIM-size slots always report fixed
"card inserted" status, even if no card is present or the SIM holder
is opened.

Probably as a side-effect of the missing card insert detection,
libccid / pcsclite are unable to detect any cards that are inserted
in the SIM-size slots after the device was attached to USB.  Only those
present at the time of the initial USB connection are reported by
pcsc_scan.