Fix a bNumberMessage issue with Cherry KC 1000 SC
The reader does not accept the value bNumberMessage = 0x00 and returns
an error. The value 0xFF (default CCID message(s)) is accepted so we use
it instead.

bNumberMessage = 0x00 is used by OpenSC for example because the device
is a keyboard and has no screen so no message to display.

Thanks to Zoltan Kelemen for the patch.