T0ProcACK(): Fix a potential integer underflow
The problem was in the funtion T0ProcACK() that is used for a Character
level reader (not TPDU or APDU reader). This is a very rare
configuration. I only have 6 such readers in my list:
https://ccid.apdu.fr/select_readers/?dwFeatures=11

The problem has been found by an automatic buzzer, not by a real problem
in the field.

Thanks to Maksim Ivanov for the bug report