Fix a potential crash in IFDHTransmitToICC()
If IFDHTransmitToICC() is used before the card is powered on then the
T=1 state machine is not yet initialized and a crash will happen when
dereferencing the checksum function pointer in t1_compute_checksum().

This will never happen when the driver is used with pcsc-lite. But it is
always safe to fix a potential bug.

The problem has been found by an automatic buzzer.

Thanks to Maksim Ivanov for the bug report