CmdPowerOn: fix a potential infinite loop
IF the reader reports bVoltageSupport = 0
AND does not support CCID_CLASS_AUTO_VOLTAGE or CCID_CLASS_AUTO_ACTIVATION
(this is bogus and only the case for the MYSMART MySMART PAD V2.0
https://ccid.apdu.fr/ccid/supported.html#0x09BE0x0002)
AND the user has conifigured ifdDriverOptions in Info.plist to use a
specific power
THEN we have an infinite loop.

Thanks to Maksim Ivanov for the bug report
"[Pcsclite-muscle] Insufficient checks in CCID"
http://lists.infradead.org/pipermail/pcsclite-muscle/2020-August/001098.html

" Hello,

The CCID free software driver is missing a few checks and graceful
handling of some error cases:

1. The device's |bVoltageSupport| having none of the low three bits
set. (The consequence of this issue is a hang in CmdPowerOn() in
"check_again: ... goto check_again". The specs say that the other bits
are reserved for future use, but theoretically this means that the
three lowest bits can be unset for some devices.) "