CmdPowerOn: fix a potention overrun with bogus ATR
If the ATR is larger than 33 bytes (MAX_ATR_SIZE) then we had a buffer
overrun.
This should never happen with correcr cards and readers.

Thanks to Maksim Ivanov for the bug report
"[Pcsclite-muscle] Insufficient checks in CCID"
http://lists.infradead.org/pipermail/pcsclite-muscle/2020-August/001098.html

" Hello,

The CCID free software driver is missing a few checks and graceful
handling of some error cases:

2. Wrong (too big) ATR size received from the device. (Would lead to a
buffer overrun in CmdPowerOn() when doing memmove() in the buffer:
https://salsa.debian.org/rousseau/CCID/-/blob/4d5cbf703c268b31c734931166c52dcb9920c0fe/src/commands.c#L294
"