CCID_Transmit: call memcpy() only with a non-NULL src
Calling memcpy(dst, NULL, 0); is not a good idea and should be avoided.
This case happens when receiving (extended) APDU in chaining mode.

hanks to Maksim Ivanov for the bug report
"[Pcsclite-muscle] Instances of Undefined behavior in CCID"
http://lists.infradead.org/pipermail/pcsclite-muscle/2020-August/001102.html

" Hello,

I found a couple of issues using the Clang's UBSan
(https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html) in the
CCID driver implementation:

2. Undefined behavior due to calling memcpy() with a null |src|
argument in CCID_Transmit(). (It's called with a zero |n| in this
case, nonetheless such calls seem to be forbidden by the C Standard.) "