Do not report an error if the wrong interface is used by the driver
With composite USB devices the pcscd daemon does not know which USB
interface correspond to the CCID reader. So all the interfaces are tried
and this generated critical errors in the logs.

Like:
00000005 readerfactory.c:1110:RFInitializeReader() Open Port 0x200000 Failed (usb:096e/080a:libudev:0:/dev/bus/usb/003/013)
00000002 readerfactory.c:375:RFAddReader() FT ePass2003Auto init failed.

When the driver reports IFD_NO_SUCH_DEVICE then the logs now have a
PCSC_LOG_INFO level so are no more displayed in red and no more present
by default.

Fixes Salsa issue #8
https://salsa.debian.org/rousseau/PCSC/-/issues/8
"Do not report an error if the wrong interface is used by the driver"