Fail if the requested protocol is not supported by reader
Some readers do not support both T=0 and T=1. So in case a protocol not
supported is requested by the application the
IFDHSetProtocolParameters() call will fail. The SCardConnect() will then
also fail with SCARD_W_UNRESPONSIVE_CARD.

A lot of tokens (with non-extractible card) do only support the protocol
of the embedded card so that is not a problem. The problem exists only
for readers where you can insert the card you want.

For example the Sysmocom sysmoOCTSIM reader does only support T=0 protocol.
Inserting and using a T=1 will confuse the reader.