Fix potential memcpy call with null dest
Add a check into CCID_Receive() in commands.c that skips calling
memcpy() in case the count is zero. This fixes the (theoretical) problem
in case both the count and the dest pointer are zero (such a call can be
made, e.g., from CmdXfrBlockAPDU_extended()) and the compiler uses the
non-null argument requirement to generate some incorrect code.

This potential issue was found by clang-tidy (the
"NonNullParamChecker" diagnostic).