Fix a possible buffer overflow in T0ProcACK
tmp_buf[] has a size of 512 bytes so the maximum length should be 0x200
(512) instead of 0x20000 (131072).

Thanks to coverity scan for the bug report.

CID 1473613 (#3 of 3): Out-of-bounds access (OVERRUN)16.
overrun-buffer-arg: Overrunning array tmp_buf of 512 bytes by passing it
to a function which accesses it at byte offset 131071 using argument
remain_len (which evaluates to 131072).