Hack to hide the ELYCTIS CNT reader reader
The ELYCTIS reader is a composite reader with 3 iterfaces:
one 0x1FFA:0x0002:ELYCTIS CL reader
and two 0x04E6:0x5810:ELYCTIS CNT reader

But the "ELYCTIS CNT reader" is in fact a "Identive CLOUD 2700 R Smart
Card Reader" renamed. They both share the same VID/PID.

Since "ELYCTIS" is before "Identive" in the alphabetical order the
VID/PID 0x04E6:0x5810 was now associated with ELYCTIS and not Identive.

The hack is to rename ELYCTIS into ~ELYCTIS for the 2 contact interfaces
so that Identive comes first in the alpahetical order.