t1_negotiate_ifsd: improve code style
Declare "retries" as an "int" instead of a "unsigned int" so the test
"if (retries <= 0)" is true for negative values of retries.

Move the test on retries _before_ sending the T=1 block. In the previous
code the block was sent but the result was ignored on the third try.

Update the check from "if (retries <= 0)" to "if (retries < 0)" so that
we really perform 3 tries instead of just 2.

Thanks to Stas Sergeev for the bug report
Fix https://github.com/LudovicRousseau/CCID/issues/91