Enable TAG_IFD_SLOT_THREAD_SAFE when supported
We only support the case of readers that can support access to all slots
at the same time. That is the case of the sysmoOCTSIM for example.

But if the reader can only support a subset of the slot at the same time
we continue as before. That is the case of the COVADIS Auriga that
declares 3 slots but can only have 2 slots active at the same time.