Multi_InterruptRead: return IFD_NO_SUCH_DEVICE when exiting
The value 0 (IFD_SUCCESS) is not a good answer when the driver is
stopping or removed.

The conversion from 0 to IFD_NO_SUCH_DEVICE is no more done in
IFDHPolling() as the comments wrongly indicates since
commit f83c1ea6806c62d39ba7ba305c92f01658796e47 (Aug 2010) but the code
for multislots was not modified at the same time.

We use IFD_NO_SUCH_DEVICE instead so that pcscd will wait a bit before
asking again.