Multi_ReadProc: wait for 100 ms in case of libusb error
We wait a bit to avoid a fast error loop.
If libusb_bulk_transfer() failed because the device has been removed
then the device will stay removed and we may fill the logs with error
messages untill the removal is detected and reported by pcscd hotplug
mechanism.

The only case in which we do NOT wait is LIBUSB_ERROR_TIMEOUT because
this case is expected and we should wait again immediatly.