Check the bSeq value when receiving a CCID frame
For multi-slots readers we now use a different USB read routine.
We can now get USB packets that were waiting in the reader but not yet
sent. For example this happens when pcscd is stopped and restarted.
On restart the driver will receive packets that were not yet read during
the previous execution of pcscd.

We must ignore packets with a bSeq that does not correspond to the bSeq
of the command sent by the reader.