Avoid potential data race for multi-slots readers
Declare "terminated" field as _Atomic since it is accessed from
different threads: one thread reads the value and the value is set from
another thread to indicate to the first thread to exit.

I have NOT found an occurence of the bug in a real execution. But the
code should be safer now.