Rewrite DEBUG_*() macros
Wrap the code in a "do { [...] } while(0)" so it is possible to use a
construct like:
  if (condition)
    DEBUG_CRITICAL("Error");
  else
    normal execution

Before the change, the DEBUG_CRITICAL() was expanded to:
  if (condition)
    if (LogLevel & DEBUG_LEVEL_CRITICAL)
      Log1(PCSC_LOG_CRITICAL, fmt)
  else
    normal execution
So the "else" was associated to the second "if" instead of the first
one.
And this is an open door for bugs.

To avoid any problem the code should have been:
  if (condition)
  {
    DEBUG_CRITICAL("Error");
  }
  else
    normal execution

DEBUG_*() maros are now safer to use.