IFDHSetProtocolParameters: always compute readTimeout
If the reader declares CCID_CLASS_AUTO_PPS_PROP the driver does not send
the PC_to_RDR_SetParameters CCID command.
But we must compute the readTimeout to use according to the card ATR.

The default read timeout of 3 seconds may not be a good value in every
cases.

The problem was detected using the SpringCard Puck One reader that does
NOT send time extension requests.