Disable USB-persist for CCID devices
The Linux kernel tries to keep the USB connection during suspend/resume
cycles. This is a bad idea if the reader/token is unpluged while the
computer is suspended.
More documentation is available at
https://docs.kernel.org/driver-api/usb/persist.html

If a CCID reader (including a composite one) is disconnected and
reconnected while the computer is suspended then the USB device will be
considered removed and connected again.
PC/SC applications will have to reconnect.

Closes: https://github.com/LudovicRousseau/PCSC/issues/154
"USB resets leave the reader connection in a broken state #154"