Suspend/resume now simulates a card removal
When a laptop (under GNU/Linux) is suspended and then resumed the libusb
interrupt read call fails with the error LIBUSB_TRANSFER_NO_DEVICE as if
the USB device had been removed.

But, because of USB-persist or something else, udev does not generate
remove/insert USB events.

The card has been powered off (during suspend) but is still present in
the reader. It is reported as present but unpowered. We need to generate
a card removal event to force pcscd and the application to negociate a
new connection to the card.

This is done by setting the bPowerFlags field to MASK_POWERFLAGS_PUP to
indicate to IFDHICCPresence() to report the card as absent on the first
call (see lines 2004-2017 of ccid_usb.c).

We now have:
[resume from suspend]
16178450 [139895155128000] ccid_usb.c:1532:InterruptRead() after (0) (5)
00000008 [139895155128000] ccid_usb.c:1544:InterruptRead() Simulate card removal
00000013 [139895155128000] ifdhandler.c:1955:IFDHICCPresence() usb:058f/9540:libudev:0:/dev/bus/usb/001/004 (lun: 0)
00000011 [139895155128000] -> 000000 65 00 00 00 00 00 09 00 00 00
00001705 [139895155128000] <- 000000 81 00 00 00 00 00 09 01 00 01
00000018 [139895155128000] ifdhandler.c:2081:IFDHICCPresence() Card absent

The card is reported as absent even if bmICCStatus=01 indicates: ICC
present and inactive (no power).

00000007 [139895155128000] eventhandler.c:358:EHStatusHandlerThread() Card Removed From Alcor Micro AU9540 00 00
00000009 [139895155128000] ifdhandler.c:318:IFDHPolling() usb:058f/9540:libudev:0:/dev/bus/usb/001/004 (lun: 0) 600000 ms
00000018 [139895155128000] ccid_usb.c:1486:InterruptRead() before (0), timeout: 600000 ms
00001706 [139895155128000] ccid_usb.c:1532:InterruptRead() after (0) (0)
00000010 [139895155128000] NotifySlotChange: 50 03
00000003 [139895155128000] ifdhandler.c:1955:IFDHICCPresence() usb:058f/9540:libudev:0:/dev/bus/usb/001/004 (lun: 0)
00000003 [139895155128000] -> 000000 65 00 00 00 00 00 0A 00 00 00
00000799 [139895155128000] <- 000000 81 00 00 00 00 00 0A 01 00 01
00000011 [139895155128000] ifdhandler.c:2081:IFDHICCPresence() Card present

Same status for bmICCStatus but this time the card is reported present.

pcscd will then generate a power up and report a SCARD_W_REMOVED_CARD
error where needed.