Revert "Suspend/resume now simulates a card removal"
This reverts commit a4401951b904707a4f93442e96ad7b137c5aa798.

We do not need this patch. We have 2 cases:
1. the card is used by an application and is powered up

After the PC resumes the card will be "present" and "unpowered" so a
card removal event will be generated because the state known by the CCID
driver was "present" and "powered".

2. the card is NOT used by an application and is not powered (because of
   pcsc-lite auto power off feature)

After the PC resumes the card will be "present" and "unpowered" (as
previously) so the CCID driver will be happy with this state and will
not signal a card removal event.
No need to report a card removal event when the card is not used by an
application.