Fix InterruptStop if called before InterruptRead
The goal is to make the polling termination (via
TAG_IFD_STOP_POLLING_THREAD) actually work even when it's
triggered before the InterruptRead has a chance to create and
publish the polling_transfer.

Without this patch, in some cases the termination was blocked for
10 minutes, because the thread that does InterruptStop might still
see the null polling_transfer meanwhile the other thread is about
to start this transfer. The proposed solution is to remember in a
separate (atomic) boolean flag whether the termination started, to
avoid this kind of race condition.