Fix possible use-after-free of polling_transfer
Guard the multi-thread usage of polling_transfer with a mutex, to
prevent the race condition between InterruptStop() and
InterruptRead(): while the former is calling
libusb_cancel_transfer(), the latter might've already called
libusb_free_transfer() on the same pointer.

This replaces the atomic variables that were previously used
for the polling_transfer pointer, because atomic variables don't
prevent this kind of use-after-free race condition.

In practice, the probability of hitting the race condition was
presumably low, but the basic scenario is the
SCardConnect/SCardDisconnect calls overlapping with the time
moments when the card is inserted/removed.