commands: CmdPowerOn: Fix nlength return value in overflow case
CmdPowerOn is supposed to return ATR and its length in
buffer/nlength.

A bounds check is performed in the code, and truncates atr_len
to the length of the received data (minus 10 bytes for the header).
However, in that case, nlength in not updated, leading to a returned
ATR with trailing garbage.

This is an edge, error case, that should never happen, and was
introduced while fixing a potentially more serious buffer overrun
in 2020.

Fixes: a382095d8b19 ("CmdPowerOn: fix a potention overrun with bogus ATR")